재능 수정 페이지 작성 완료
재능 수정 클릭 시 재능번호에 맞는 데이터 불러오고 등록시 DB UPDATE 작동 확인

+++ jPanda/src/main/java/com/kakao/jPanda/bsm/domain/Talent.java
@@ -6,16 +6,16 @@ import lombok.Data;
 public class Talent {
 	private Long 	talentNo;
 	private String 	sellerId;
-	private int 	upperCategoryNo;
-	private int 	lowerCategoryNo;
+	private Long 	upperCategoryNo;
+	private Long 	lowerCategoryNo;
 	private String 	mainImg;
 	private String 	title;
 	private String  content;
-	private int  	bamboo;
-	private int  	saleBamboo;
+	private Long  	bamboo;
+	private Long  	saleBamboo;
 	private String 	summary;
 	private String 	status;
-	private int  	viewCount;
+	private Long  	viewCount;
 	private String 	regDate;
 	private String 	statusDate;
 }
